separate the MainFloatingState interface
